Fix Signed Addition

+++ Ashhar/CNN_using_C_fwd_fixedpt/matrix.h
@@ -2,5 +2,5 @@ struct Matrix
 {
     int rows;
     int columns;
-    float **elements;
+    int **elements;
 };
